pip_requirements.txt
certifi==2022.12.7
charset-normalizer==2.1.1
cycler==0.11.0
faiss-gpu==1.7.2
filelock==3.12.4
fonttools==4.38.0
fsspec==2023.10.0
h5py==3.8.0
huggingface-hub==0.18.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.40.1
matplotlib==3.5.3
numba==0.57.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==22.0
pandas==2.0.3
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
safetensors==0.4.0
scikit-dimension==0.3.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
timm==0.9.8
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.13
